Adds script to generate 10 million business and then names it appropriately
